chore(infra): change entrypoint command
Removes the rails server command from the entrypoint and add it to the compose spec, this should enable to run docker-compose run commands on the app service
Fix bootsnap cache dir with env variable.
Removes the gem cache volume, since we are installing gems frequently and becomes annoying to clear the volumes each time.