feat(account): creates account model and specs
Creates the account model, it is annotated with it fields by the annotate gem.
The account model represents an account of the system, it has a balance, as a monetized field by the money gem.
Account has a json representation through the as_json method, we should evolve to serializers if this method becomes too complex