feat(accounts): add actions to accounts controller
Adds the create and show actions to the accounts controller V1.
The create actions is not protected to make it possible to unauthenticated users create an account.
The show method is protected.
Create method uses the create account service.
Specs tests basically the controller logic, it does not test if services behave as expected, for instance.
The service test should be enough to cover it.